Extract generic DOM helpers from options.ts
